listening-comp/requirements.txt
chromadb
streamlit
boto3
youtube_transcript_api
python-dotenv
fastapi
uvicorn
pyyaml
requests